+++ src/librtprocess/src/preprocess/CA_correct.cc
@@ -215,7 +215,9 @@ rpError CA_correct(
     float *RawDataTmp = Gtmp + (winh * (winw + (winw & 1))) / 2;
     //block CA shift values and weight assigned to block
     float *const blockwt = Gtmp + height * width;
-    memset(blockwt, 0, vblsz * hblsz * (2 * 2 + 1) * sizeof(float));
+    memset(blockwt,
+           0,
+           static_cast<size_t>(vblsz) * static_cast<size_t>(hblsz) * (2u * 2u + 1u) * sizeof(float));
     float (*blockshifts)[2][2] = (float (*)[2][2])(blockwt + vblsz * hblsz);
 
     // Because we can't break parallel processing, we need a switch do handle the errors
